refactor(mvc): rename scheduler module to dispatch
